+++ include/libfly/env/geometry.hpp
@@ -27,12 +27,13 @@
 #include <vector>
 
 #include "libfly/system/VoS.hpp"
+#include "libfly/system/property.hpp"
 #include "libfly/utility/core.hpp"
 
 /**
  * \file geometry.hpp
  *
- * @brief ...
+ * @brief Utilities for manipulating local environments of atoms.
  *
  */
 
@@ -48,6 +49,8 @@ namespace fly::env {
    *    \frac{1}{N} \sum_{i=1}^{N}{x_i}
    *
    * \endrst
+   *
+   * @param x Set of atoms with the Position property.
    */
   template <typename... M>
   auto centroid(system::VoS<M...> const &x) noexcept -> Vec {
@@ -71,9 +74,13 @@ namespace fly::env {
    * this is:
    *
    * .. math::
-   *    \Delta_\text{RMSD} = \sqrt{\sum_{i=1}^N{\left\lVert Mx_i - y_i \right\rVert^2}}
+   *    \Delta_\text{GRMSD} = \sqrt{\sum_{i=1}^N{\left\lVert Mx_i - y_i \right\rVert^2}}
    *
    * \endrst
+   *
+   * @param x Set of atoms with the Position property.
+   * @param y Set of atoms with the Position property.
+   * @param M Transformation matrix.
    */
   template <typename E, typename... M1, typename... M2>
   auto grmsd(E const &M, system::VoS<M1...> const &x, system::VoS<M2...> const &y) -> double {
@@ -98,9 +105,12 @@ namespace fly::env {
    * atoms, for two sets of atoms with positions :math:`x_i` and :math:`y_i` this is:
    *
    * .. math::
-   *    \Delta_\text{GRMSD} = \sqrt{\sum_{i=1}^N{\left\lVert x_i - y_i \right\rVert^2}}
+   *    \Delta_\text{RMSD} = \sqrt{\sum_{i=1}^N{\left\lVert x_i - y_i \right\rVert^2}}
    *
    * \endrst
+   *
+   * @param x Set of atoms with the Position property.
+   * @param y Set of atoms with the Position property.
    */
   template <typename... M1, typename... M2>
   auto rmsd(system::VoS<M1...> const &x, system::VoS<M2...> const &y) -> double {
@@ -122,6 +132,9 @@ namespace fly::env {
    * \endrst
    *
    * Uses the "Kabsch algorithm" see: https://en.wikipedia.org/wiki/Kabsch_algorithm
+   *
+   * @param x Set of atoms with the Position property.
+   * @param y Set of atoms with the Position property.
    */
   template <typename... M1, typename... M2>
   auto ortho_onto(system::VoS<M1...> const &x, system::VoS<M2...> const &y) -> Mat {
@@ -129,9 +142,6 @@ namespace fly::env {
     Mat H = Mat::Zero();
 
     ASSERT(x.size() == y.size(), "Sizes must match, {}!={}!", x.size(), y.size());
-
-    // ASSERT(near(gnorm(centroid(x)), 0.0), "Centroid of x={}", centroid(x));
-    // ASSERT(near(gnorm(centroid(y)), 0.0), "Centroid of y={}", centroid(y));
 
     for (int i = 0; i < x.size(); i++) {
       H.noalias() += x[i][r_] * y[i][r_].transpose();
@@ -144,179 +154,186 @@ namespace fly::env {
     return svd.matrixV() * svd.matrixU().transpose();
   }
 
-  //   namespace detail {
+  namespace detail {
 
-  //     /** @brief The maximum number of atoms that can lie in the same plane. */
-  //     inline constexpr std::size_t MAX_COPLANAR_ATOMS = 10;
+    /** @brief The maximum number of atoms that can lie in the same plane. */
+    inline constexpr int MAX_COPLANAR_ATOMS = 10;
 
-  //     /**
-  //      * @brief Helper function, verifies addition of n^th atom matches all previous atoms.
-  //      */
-  //     template <typename... M1, typename... M2>
-  //     bool within_tol_up_to(system::VoS<M1...> const &mut, system::VoS<M2...> const &ref, double tol, std::size_t n) {
-  //       //
-  //       for (std::size_t i = 0; i < std::min(n, MAX_COPLANAR_ATOMS); ++i) {
-  //         // Intra atomic distances
-  //         double ref_ni = norm(ref[n][r_] - ref[i][r_]);
-  //         double mut_ni = norm(mut[n][r_] - mut[i][r_]);
+    /**
+     * @brief Helper function, verifies addition of n^th atom matches all previous atoms.
+     */
+    template <typename... M1, typename... M2>
+    bool within_tol_up_to(system::VoS<M1...> const &mut, system::VoS<M2...> const &ref, double tol, int n) {
+      //
+      for (int i = 0; i < std::min(n, MAX_COPLANAR_ATOMS); ++i) {
+        // Intra atomic distances
+        double ref_ni = gnorm(ref[n][r_] - ref[i][r_]);
+        double mut_ni = gnorm(mut[n][r_] - mut[i][r_]);
 
-  //         if (std::abs(ref_ni - mut_ni) > tol) {
-  //           return false;
-  //         }
-  //       }
-  //       return true;
-  //     }
+        if (std::abs(ref_ni - mut_ni) > tol) {
+          return false;
+        }
+      }
+      return true;
+    }
 
-  //     /**
-  //      * @brief Recursive implementation of for_equiv_perms.
-  //      *
-  //      * Uses the GREEDY_PERM algorithm, recursively finds a new atom, such that the order that
-  //      * the intra-atom distances in each system::VoS match (within tolerance SQRT_2 * delta). The
-  //      * final permutation must be able to be transformed (via ortho_onto(mut, other)) such that the
-  //      * RMSD between the two sets is less than delta.
-  //      */
-  //     template <typename... M1, typename... M2, typename F>
-  //     bool for_equiv_perms_impl(system::VoS<M1...> &mut, system::VoS<M2...> const &ref, double delta, std::size_t n, F const &f) {
-  //       // Termination criterion.
-  //       if (n >= mut.size()) {
-  //         //
-  //         Mat O = ortho_onto(mut, ref);
+    /**
+     * @brief Recursive implementation of for_equiv_perms.
+     *
+     * Uses the GREEDY_PERM algorithm, recursively finds a new atom, such that the order that
+     * the intra-atom distances in each system::VoS match (within tolerance SQRT_2 * delta). The
+     * final permutation must be able to be transformed (via ortho_onto(mut, other)) such that the
+     * RMSD between the two sets is less than delta.
+     */
+    template <typename... M1, typename... M2, typename F>
+    bool for_equiv_perms_impl(system::VoS<M1...> &mut, system::VoS<M2...> const &ref, double delta, int n, F const &f) {
+      // Termination criterion.
+      if (n >= mut.size()) {
+        //
+        Mat O = ortho_onto(mut, ref);
 
-  //         if (double dr = rmsd(O, mut, ref); dr < delta) {
-  //           return std::invoke(f, O, dr);
-  //         } else {
-  //           return false;
-  //         }
-  //       }
+        if (double dr = grmsd(O, mut, ref); dr < delta) {
+          return std::invoke(f, O, dr);
+        } else {
+          return false;
+        }
+      }
 
-  //       using std::swap;  // ADL
+      using std::swap;  // ADL
 
-  //       // Attempt to find an atom to put in i^th position.
-  //       for (std::size_t i = n; i < ref.size(); ++i) {
-  //         if (mut[i](Colour{}) == ref[n](Colour{})) {
-  //           // Test next candidate
-  //           swap(mut[n], mut[i]);
+      // Attempt to find an atom to put in i^th position.
+      for (int i = n; i < ref.size(); ++i) {
+        if (mut[i][col_] == ref[n][col_]) {
+          // Test next candidate
+          swap(mut[n], mut[i]);
 
-  //           // Verify all other distances then recurse
-  //           if (within_tol_up_to(mut, ref, delta * M_SQRT2, n)) {
-  //             if (for_equiv_perms_impl(mut, ref, delta, n + 1, f)) {
-  //               // Enable early exit.
-  //               return true;
-  //             }
-  //           }
+          // Verify all other distances then recurse
+          if (within_tol_up_to(mut, ref, delta * M_SQRT2, n) && for_equiv_perms_impl(mut, ref, delta, n + 1, f)) {
+            // Enable early exit.
+            return true;
+          }
 
-  //           // Undo swap such that if we have to back-track order is the same.
-  //           swap(mut[n], mut[i]);
-  //         }
-  //       }
+          // Undo swap such that if we have to back-track order is the same.
+          swap(mut[n], mut[i]);
+        }
+      }
 
-  //       return false;
-  //     }
+      return false;
+    }
 
-  //   }  // namespace detail
+  }  // namespace detail
 
-  //   /**
-  //    * @brief Create a function object to explore equivalent permutations of another system::VoS.
-  //    *
-  //    * This function is for greedily exploring the permutations of atoms in another system::VoS such
-  //    * that, after an orthogonal transformation generated by ortho_onto(), the RMSD between the two
-  //    * sets of atoms is less than delta.
-  //    *
-  //    * Example:
-  //    *
-  //    * @code{.cpp}
-  //    *
-  //    * for_equiv_perms(ref, 0.2)(1, mut, [](Mat const & O, double RMSD){
-  //    *     // If this function is called then "mut" has been permuted into an equivalent permutation.
-  //    *     // "O" is the matrix that maps "mut" to "ref" such that that:
-  //    *     //     "RMSD" == RMSD(O, mut, ref) and "RMSD" < delta.
-  //    *
-  //    *     // We could now do something with "RMSD" and "O". We must not mutate "ref" or "mut".
-  //    *
-  //    *     // If we "return true" then exploration/function terminates.
-  //    *     // If we "return false" then exploration/function continues.
-  //    * });
-  //    *
-  //    * @endcode
-  //    *
-  //    * for_equiv_perms will not permute the first "n" (== 1 in the above example) atoms in "mut". This
-  //    * is useful if there exist a number of atoms who's permutation is known.
-  //    *
-  //    */
-  //   template <typename... M>
-  //   auto for_equiv_perms(system::VoS<M...> const &ref, double delta) {
-  //     return [&ref, delta](std::size_t n, auto &mut, auto &&callback) {
-  //       ASSERT(ref.size() == mut.size(), "Sizes must match!");
+  /**
+   * @brief Explore equivalent permutations of atoms in two sets.
+   *
+   * This function is for greedily exploring the permutations of atoms in ``mut`` such that, after an orthogonal transformation
+   * generated by ``ortho_onto()``, the ``grmsd()`` between the two sets of atoms is less than ``delta``.
+   *
+   * \rst
+   *
+   * Example:
+   *
+   * .. include:: ../../examples/env/geometry.cpp
+   *    :code:
+   *
+   * \endrst
+   *
+   * ``for_equiv_perms()`` will not permute the first ``n`` (== 1 in the above example) atoms in ``mut``. This
+   * is useful if there exist a number of atoms who's permutation is known.
+   *
+   *
+   * @param mut Set of atoms with the Position & Colour properties that will be permuted to match ``ref``.
+   * @param ref The reference set of atoms with the Position & Colour properties.
+   * @param delta The maximum ``grmsd()`` between the permuted ``mut`` and ``ref``
+   * @param n The first ``n`` atoms in ``mut`` will not be permuted.
+   * @param f An invokable invoked with signature ``(fly::Mat const& O, double rmsd) -> bool`` called for each permutation of ``mut``
+   * that matches ``ref``, see the example above for details.
+   */
+  template <typename... M1, typename... M2, typename F>
+  auto for_equiv_perms(system::VoS<M1...> &mut, system::VoS<M2...> const &ref, double delta, int n, F const &f) -> void {
+    //
+    ASSERT(ref.size() == mut.size(), "Sizes must match, {}!={}!", ref.size(), mut.size());
 
-  //       if constexpr (std::is_same_v<std::remove_cv_t<decltype(mut)>, system::VoS<M...>>) {
-  //         ASSERT(&ref != &mut, "Cannot perm onto self.");
-  //       }
+    if constexpr (std::is_same_v<system::VoS<M1...>, system::VoS<M2...>>) {
+      ASSERT(&ref != &mut, "Cannot perm onto self.", 0);
+    }
 
-  //       detail::for_equiv_perms_impl(mut, ref, delta, n, std::forward<decltype(callback)>(callback));
-  //     };
-  //   }
+    detail::for_equiv_perms_impl(mut, ref, delta, n, f);
+  }
 
-  //   /**
-  //    * @brief A geometry models a local distribution of periodically resolved Atoms.
-  //    *
-  //    * The distribution of atoms is centred on the first atom.
-  //    */
-  //   template <typename... Mems>
-  //   class Geometry : public system::VoS<Mems...> {
-  //   public:
-  //     /**
-  //      * @brief Returned by permute_onto(delta, x), contains extra info about the permutation.
-  //      */
-  //     struct PermResult {
-  //       /** @brief  Orthogonal transformation required to map this onto x. */
-  //       Mat O;
-  //       /** @brief Equal to rmsd(O, *this, x). */
-  //       double rmsd;
-  //     };
+  /**
+   * @brief A Geometry models a local distribution of periodically resolved Atoms.
+   *
+   * The distribution of atoms is centred on the first atom, i.e. the first atom cannot move and is always fixed during permutation
+   * operations.
+   *
+   * A Geometry is a ``fly::system::VoS``  with the default properties Position and Colour.
+   */
+  template <typename... Pr>
+  class Geometry : public system::VoS<Position, Colour, Pr...> {
+  public:
+    /**
+     * @brief Returned by permutation methods, contains extra info about the permutation.
+     */
+    struct GeoInfo {
+      Mat O;        ///< Orthogonal transformation required to map this onto ``x``.
+      double rmsd;  ///<  Equal to ``grmsd(O, *this, x)``.
+    };
 
-  //     /**
-  //      * @brief Attempt to permute the atoms in this Geometry into the same order as the atoms in
-  //      * other.
-  //      *
-  //      * @return std::optional<PermResult> An engaged PermResult if a permutation was found.
-  //      */
-  //     template <typename... Ts>
-  //     std::optional<PermResult> permute_onto(system::VoS<Ts...> const &other, double delta) {
-  //       std::optional<PermResult> res = std::nullopt;
+    /**
+     * @brief Find the first permutation of the atoms in this Geometry onto ``other`` within tolerance ``delta``.
+     *
+     * @param other The reference set of atoms with the Position & Colour properties.
+     * @param delta The maximum ``grmsd()`` between the permuted ``this`` and ``ref`` if this operation succeeds.
+     * @return std::optional<GeoInfo> An engaged ``GeoInfo`` if a permutation was found.
+     */
+    template <typename... Ts>
+    std::optional<GeoInfo> permute_onto(system::VoS<Ts...> const &other, double delta) {
+      std::optional<GeoInfo> res = std::nullopt;
 
-  //       for_equiv_perms(other, delta)(1, *this, [&](Mat const &O, double rmsd) {
-  //         res = PermResult{O, rmsd};
-  //         // First match accepted
-  //         return true;
-  //       });
+      for_equiv_perms(other, delta, 1, *this, [&res](Mat const &O, double rmsd) {
+        res = GeoInfo{O, rmsd};
+        return true;  // First match accepted
+      });
 
-  //       return res;
-  //     }
+      return res;
+    }
 
-  //     /**
-  //      * @brief Find the best permutation of the atoms in this geomety onto other
-  //      *
-  //      * @return std::optional<PermResult> An engaged PermResult if a permutation was found.
-  //      */
-  //     template <typename... Ts>
-  //     std::optional<PermResult> best_perm_onto(system::VoS<Ts...> const &other, double delta, Geometry<Mems...> &scratch) {
-  //       std::optional<PermResult> res = std::nullopt;
+    /**
+     * @brief Find the best permutation of the atoms in this Geometry onto ``other``.
+     *
+     * @param other The reference set of atoms with the Position & Colour properties.
+     * @param delta The maximum ``grmsd()`` between the permuted ``this`` and ``ref`` for this operation to succeed.
+     * @param scratch_best An optional pointer to scratch space that the algorithm can use (otherwise this function will allocate).
+     * @return std::optional<GeoInfo> An engaged ``GeoInfo`` if a permutation was found.
+     */
+    template <typename... Ts>
+    std::optional<GeoInfo> best_perm_onto(system::VoS<Ts...> const &other, double delta, Geometry *scratch_best = nullptr) {
+      //
+      std::optional<GeoInfo> res = std::nullopt;
 
-  //       for_equiv_perms(other, delta)(1, *this, [&](Mat const &O, double rmsd) {
-  //         if (!res || res->rmsd < rmsd) {
-  //           res = PermResult{O, rmsd};
-  //           scratch = *this;
-  //         }
+      Geometry tmp;
 
-  //         return false;
-  //       });
+      if (!scratch_best) {
+        scratch_best = &tmp;
+      }
 
-  //       if (res) {
-  //         *this = scratch;
-  //       }
+      for_equiv_perms(other, delta)(1, *this, [&](Mat const &O, double rmsd) {
+        if (!res || res->rmsd < rmsd) {
+          res = GeoInfo{O, rmsd};  // Cache best info.
+          scratch_best = *this;    // Copy best so far into scratch.
+        }
 
-  //       return res;
-  //     }
-  //   };
+        return false;  // Explore all.
+      });
+
+      if (res) {
+        using std::swap;
+        swap(*this, scratch_best);
+      }
+
+      return res;
+    }
+  };
 
 }  // namespace fly::env